Move scroll function into separate file.
